Refactor OpenVPN view
- Move import logic to external modifier
- Fix unsorted OpenVPN rows (pull, redirect gw)
- Move preview builder to separate file
